+++ src/Estimator.php
@@ -4,7 +4,7 @@ namespace Rubix\ML;
 
 use Rubix\ML\Datasets\Dataset;
 
-interface Estimator extends Model
+interface Estimator
 {
     /**
      * Return the estimator type.
@@ -12,6 +12,20 @@ interface Estimator extends Model
      * @return \Rubix\ML\EstimatorType
      */
     public function type() : EstimatorType;
+
+    /**
+     * Return the data types that the model is compatible with.
+     *
+     * @return \Rubix\ML\DataType[]
+     */
+    public function compatibility() : array;
+
+    /**
+     * Return the settings of the hyper-parameters in an associative array.
+     *
+     * @return mixed[]
+     */
+    public function params() : array;
 
     /**
      * Make predictions from a dataset.