+++ src/Transformers/ImageVectorizer.php
@@ -74,7 +74,7 @@ class ImageVectorizer implements Transformer, Stateful
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      * @throws \InvalidArgumentException