+++ src/Transformers/IntervalDiscretizer.php
@@ -114,7 +114,7 @@ class IntervalDiscretizer implements Transformer, Stateful
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      * @throws \InvalidArgumentException