+++ src/Transformers/KNNImputer.php
@@ -127,7 +127,7 @@ class KNNImputer implements Transformer, Stateful
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      * @throws \RuntimeException