+++ src/Transformers/LinearDiscriminantAnalysis.php
@@ -131,7 +131,7 @@ class LinearDiscriminantAnalysis implements Transformer, Stateful
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      * @throws \InvalidArgumentException