+++ src/Transformers/MaxAbsoluteScaler.php
@@ -59,7 +59,7 @@ class MaxAbsoluteScaler implements Transformer, Stateful, Elastic
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      */