+++ src/Transformers/MissingDataImputer.php
@@ -98,7 +98,7 @@ class MissingDataImputer implements Transformer, Stateful
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      * @throws \InvalidArgumentException