+++ src/Transformers/OneHotEncoder.php
@@ -65,7 +65,7 @@ class OneHotEncoder implements Transformer, Stateful
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      */