+++ src/Transformers/PrincipalComponentAnalysis.php
@@ -142,7 +142,7 @@ class PrincipalComponentAnalysis implements Transformer, Stateful
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      * @throws \InvalidArgumentException