+++ src/Transformers/RobustStandardizer.php
@@ -97,7 +97,7 @@ class RobustStandardizer implements Transformer, Stateful
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      */