+++ src/Transformers/TfIdfTransformer.php
@@ -86,7 +86,7 @@ class TfIdfTransformer implements Transformer, Stateful, Elastic
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      * @throws \InvalidArgumentException