+++ src/Transformers/VarianceThresholdFilter.php
@@ -91,7 +91,7 @@ class VarianceThresholdFilter implements Transformer, Stateful
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      */