+++ src/Transformers/WordCountVectorizer.php
@@ -113,7 +113,7 @@ class WordCountVectorizer implements Transformer, Stateful
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      */