+++ src/Transformers/ZScaleStandardizer.php
@@ -123,7 +123,7 @@ class ZScaleStandardizer implements Transformer, Stateful, Elastic
     }
 
     /**
-     * Fit the transformer to the dataset.
+     * Fit the transformer to a dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      */